working on content pages

+++ themes/default/views/oc-panel/pages/content/list.php
@@ -14,6 +14,7 @@
     rel="tooltip" title="<?=__('Edit')?>">
     <?=__('New')?>
 </a>
+
 <?= FORM::open(Route::url('oc-panel',array('controller'=>'content','action'=>'list')), array('method'=>'GET','class'=>'form-horizontal', 'id'=>'locale_form','enctype'=>'multipart/form-data'))?>
     <div class="form-group">
         
@@ -37,28 +38,29 @@
         <th></th>
     </tr>
 <?foreach ($contents as $content):?>
- 
-    <tr>
-        <td><?=$content->title?></td>
-        <td><?=$content->locale?></td>
-        <td><?=$content->created?></td>
-        <td><?=$content->seotitle?></td>
-        <td><?=$content->status?></td>
-        <td width="5%">
-            
-            <a class="btn btn-primary" 
-                href="<?=Route::url('oc-panel', array('controller'=>'content','action'=>'edit','id'=>$content))?>" 
-                rel="tooltip" title="<?=__('Edit')?>">
-                <i class="glyphicon   glyphicon-edit"></i>
-            </a>
-            <a class="btn btn-danger" 
-                href="<?=Route::url('oc-panel', array('controller'=>'content','action'=>'delete','id'=>$content))?>" 
-                rel="tooltip" title="<?=__('Delete')?>">
-                <i class="glyphicon   glyphicon-trash"></i>
-            </a>
+    <?if(isset($content->title)):?>
+        <tr>
+            <td><?=$content->title?></td>
+            <td><?=$content->locale?></td>
+            <td><?=$content->created?></td>
+            <td><?=$content->seotitle?></td>
+            <td><?=$content->status?></td>
+            <td width="5%">
+                
+                <a class="btn btn-primary" 
+                    href="<?=Route::url('oc-panel', array('controller'=>'content','action'=>'edit','id'=>$content))?>" 
+                    rel="tooltip" title="<?=__('Edit')?>">
+                    <i class="glyphicon   glyphicon-edit"></i>
+                </a>
+                <a class="btn btn-danger" 
+                    href="<?=Route::url('oc-panel', array('controller'=>'content','action'=>'delete','id'=>$content))?>" 
+                    rel="tooltip" title="<?=__('Delete')?>">
+                    <i class="glyphicon   glyphicon-trash"></i>
+                </a>
 
-        </td>
-    </tr>
+            </td>
+        </tr>
+    <?endif?>
     
 <?endforeach?>
 </table>
